requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate 
psycopg2-binary
python-dotenv
Flask-Cors